+++ src/test/kotlin/features/glue/StepDefs.kt
@@ -17,9 +17,11 @@ import com.jetbrains.python.fixtures.PyLightProjectDescriptor
 import com.jetbrains.python.psi.LanguageLevel
 import com.jetbrains.python.psi.PyFile
 import com.jetbrains.snakecharm.SnakemakeTestUtil
+import com.jetbrains.snakecharm.framework.SmkFrameworkDeprecationProvider
 import com.jetbrains.snakecharm.framework.SmkSupportProjectSettings
 import io.cucumber.java.en.Given
 import javax.swing.SwingUtilities
+import kotlin.Throws
 import kotlin.test.fail
 
 
@@ -131,6 +133,12 @@ class StepDefs {
         // XXX: Post Startup activities should end before this if everything goes OK
         SnakemakeWorld.myInjectionFixture = InjectionTestFixture(SnakemakeWorld.fixture())
 
+        // XXX: reset Snakemake API settings if smth was overridden in other tests
+        ApplicationManager.getApplication().invokeAndWait {
+            ApplicationManager.getApplication().runWriteAction {
+                SmkFrameworkDeprecationProvider.getInstance().reinitializeInTests()
+            }
+        }
         setProjectSdk("python with snakemake")
 
         if (projectType != "snakemake with disabled framework") {